refactor: update login page layout and styles
