+++ frontend/src/app/auth/login/page.tsx
@@ -1,8 +1,7 @@
+import { getAppSession } from "@/app/actions";
 import LoginForm from "@/lib/components/auth/LoginForm";
-import { redirect } from 'next/navigation';
-import { getAppSession } from '@/app/actions';
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/lib/components/common/Card'
-import { getTranslations } from 'next-intl/server'
+import { getTranslations } from "next-intl/server";
+import { redirect } from "next/navigation";
 
 export default async function Login() {
     const { sessionData } = await getAppSession();
@@ -13,18 +12,16 @@ export default async function Login() {
 
     const t = await getTranslations("Sidebar");
 
-    return (<>
-        <div className='w-1/3 space-y-4 mx-auto py-32'>
-            <Card>
-                <CardHeader>
-                    <CardTitle>
-                        {t('creator-login')}
-                    </CardTitle>
-                </CardHeader>
-                <CardContent>
+    return (
+        <>
+            <div className="flex flex-col justify-center items-center h-screen bg-gradient-to-b from-primary/10">
+                <div className="max-w-[450px] w-full p-5">
+                    <h1 className="text-3xl font-semibold mb-5">
+                        Welcome back.
+                    </h1>
                     <LoginForm />
-                </CardContent>
-            </Card>
-        </div>
-    </>);
+                </div>
+            </div>
+        </>
+    );
 }
